borrow book  done

+++ src/app/modules/user/user.service.ts
@@ -1,6 +1,8 @@
 import { TMember, TMemberUpdate } from "./user.interface";
 import prisma from "../../../shared/prisma";
 import { Prisma } from "@prisma/client";
+import ApiError from "../../errors/ApiError";
+import httpStatus from "http-status";
 
 // Define createBook function to insert a new book into the database
 const createUser = async (
@@ -14,6 +16,7 @@ const createUser = async (
 
 const getAllUser = async () => {
   const users = await prisma.member.findMany();
+
   return users;
 };
 
@@ -21,6 +24,9 @@ const getSingleUser = async (UserId: string) => {
   const user = await prisma.member.findUnique({
     where: { memberId: UserId },
   });
+  if (!user) {
+    return new ApiError(httpStatus.NOT_FOUND, "User Not found");
+  }
   return user;
 };
 
@@ -29,6 +35,7 @@ const updateUser = async (
   payload: Partial<TMemberUpdate>
 ): Promise<TMember> => {
   const { name, email, phone } = payload;
+
   const updateUser = await prisma.member.update({
     where: { memberId: UserId },
     data: {
@@ -41,6 +48,12 @@ const updateUser = async (
 };
 
 const deleteUser = async (UserId: string) => {
+  const user = await prisma.member.findUnique({
+    where: { memberId: UserId },
+  });
+  if (!user) {
+    return new ApiError(httpStatus.NOT_FOUND, "User Not found");
+  }
   const result = await prisma.member.delete({
     where: { memberId: UserId },
   });
